requirements.txt
pandas>=2.0.3
numpy>=1.24.3
openpyxl>=3.1.2
torch>=2.0.1
torchvision>=0.15.2
opencv-python>=4.8.0.74
Pillow>=10.0.0
requests>=2.31.0
scikit-learn>=1.3.0
xgboost>=1.7.6
matplotlib>=3.7.2
seaborn>=0.12.2
plotly>=5.15.0
geopandas>=0.13.2
folium>=0.14.0
tqdm>=4.65.0
python-dotenv>=1.0.0
